backend/requirements.txt
# ADAS Backend - Windows Server Python 3.11
# ALL packages use pre-built wheels - NO compilation needed

# ================================================
# Core Framework
# ================================================
fastapi==0.115.0
uvicorn==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.1
python-multipart==0.0.12

# ================================================
# Computer Vision & AI (Pre-built for Windows)
# ================================================
opencv-python==4.10.0.84
numpy==1.26.4
mediapipe==0.10.14

# PyTorch CPU (Let pip resolve compatible versions)
--extra-index-url https://download.pytorch.org/whl/cpu
torch>=2.0.0,<2.4.0
torchvision>=0.15.0,<0.19.0

# YOLOv11 (will auto-select compatible torch version)
ultralytics>=8.0.0

# ================================================
# Utilities
# ================================================
python-dotenv==1.0.1
httpx==0.27.2
aiofiles==24.1.0
pillow==10.4.0

# ================================================
# Authentication & Security (Phase 8)
# ================================================
python-jose[cryptography]==3.3.0
bcrypt==4.2.0

# ================================================
# Text-to-Speech (Phase 6)
# ================================================
gTTS==2.5.3

# ================================================
# Object Tracking (ByteTrack)
# ================================================
filterpy==1.4.5
scipy==1.11.4

# ================================================
# System Monitoring (Phase 9)
# ================================================
psutil==6.1.0

# ================================================
# Testing (Optional)
# ================================================
pytest==8.3.0
